pyproject.toml
[project]
name = "browser-agent"
version = "0.1.0"
description = "Ghost Browser Agent - undetectable browser automation via Chrome Extension + WebSocket"
requires-python = ">=3.11"
dependencies = [
    "websockets>=13.0",
    "openai>=1.0",
    "Pillow>=10.0",
    "pyyaml>=6.0",
    "python-telegram-bot>=21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
